Update the 3.13 plan (#590)
This builds on Mark's excellent overview.

1. I broke out copy-and-patch into its own thing -- I think it's big
enough and parts of it (especially the build-time stuff), is
parallelizable.
2. I added PEP 554 (subinterpreters in the stdlib).
3. For the detailed planning parts, I've created METABUG issues where we
can link to ideas and/or cpython issues related to each step. I find
issue links showing the status (open/closed etc.) of the issue extremely
useful for tracking progress, but those only work issue-to-issue, not in
a file or wiki.
4. For full transparency, I'm debating how much to prioritize the memory
management work (at the expense of other things). I'm leaving it here
for now, but I think mainly to see if there are quick cheap experiments
we can run to generate more concrete solutions.

---------

Co-authored-by: Brandt Bucher <brandtbucher@gmail.com>